pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
name = "quadrantrecon"
version = "1.0.1"
description = "Image recognition software to crop photos of marine ecosystems. Made for the MBOM Pole to Pole project."
authors = [
    {name = "Lino Bigatti", email = "linobigatti@protonmail.com"},
]
dependencies = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "attrs>=23.2.0",
    "autocommand>=2.2.2",
    "cffi>=1.16.0",
    "charset-normalizer>=3.3.2",
    "contourpy>=1.2.1",
    "cryptography>=42.0.5",
    "cycler>=0.12.1",
    "fastjsonschema>=2.19.1",
    "filelock>=3.14.0",
    "fonttools>=4.51.0",
    "fsspec>=2024.3.1",
    "future>=1.0.0",
    "gmpy2>=2.2.0a1",
    "idna>=3.6",
    "inflect>=7.2.1",
    "jaraco-context>=4.3.0",
    "jaraco-functools>=4.0.0",
    "jaraco-text>=3.12.0",
    "Jinja2>=3.1.3",
    "kiwisolver>=1.4.5",
    "MarkupSafe>=2.1.5",
    "matplotlib>=3.8.4",
    "more-itertools>=10.2.0",
    "mpmath>=1.3.0",
    "networkx>=3.3",
    "numpy==1.26.4",
    "opencv-python>=4.9.0.80",
    "ordered-set>=4.1.0",
    "packaging>=24.0",
    "pandas>=1.5.3",
    "pillow>=10.3.0",
    "platformdirs>=4.2.0",
    "pooch>=1.8.1",
    "pybind11>=2.12.0",
    "pycparser>=2.22",
    "pyparsing>=3.1.2",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2024.1",
    "PyYAML>=6.0.1",
    "requests>=2.31.0",
    "scipy>=1.13.0",
    "setuptools>=69.0.3",
    "six>=1.16.0",
    "smbus>=1.1",
    "sympy>=1.12",
    "TBB>=0.2",
    "tomli>=2.0.1",
    "trove-classifiers>=2024.4.10",
    "typeguard>=4.2.1",
    "typing-extensions>=4.11.0",
    "urllib3>=1.26.18",
    "validate-pyproject>=0.16",
    "wheel>=0.43.0",
    "tk>=0.1.0",
    "piexif>=1.1.2",
    "tqdm>=4.67.1",
    "segment-anything-py>=1.0.1",
    #"segment_anything @ file:///${PROJECT_ROOT}/src/sam/dist/segment_anything-1.0-py3-none-any.whl",
    "iptcinfo3>=2.1.4"
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Environment :: GPU :: NVIDIA CUDA :: 12",
    "Framework :: Matplotlib",
    "Natural Language :: English",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Image Processing",
]

[tool.pdm]
distribution = true
